requirements.txt
langchain
langchain_core
langchain_community
openai
python-dotenv
sqlalchemy
langgraph
pydantic
fastapi
langchain_openai